training-studio/backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# AI/ML - Transcription
openai-whisper>=20231117

# AI/ML - Speaker Diarization
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio Processing - Prosody
librosa>=0.10.1
parselmouth-praat>=0.4.3
numpy>=1.26.0
scipy>=1.11.0
soundfile>=0.12.1

# Video/Image Processing - Facial
mediapipe>=0.10.9
opencv-python>=4.9.0

# DeepFace for emotion recognition (optional, heavier)
# deepface>=0.0.79
# Lighter alternative: py-feat
py-feat>=0.6.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
anthropic>=0.18.0

# Progress tracking
tqdm>=4.66.0
